Added some juice to platforms, may need to do some speed adjustments

+++ SFML_Platformer/PlayerHero.h
@@ -11,6 +11,6 @@ public:
 
 private:
 	
-	int sensitivity; //sets acceleration control for movement
+	float sensitivity; //sets acceleration control for movement
 
 };
